Change logo dimensions to be divisible by 2
The logo was 71x62 and resized to be 35x31.
As the ratio is slightly off because 71 is not divisible by 2
(there are no half pixels), this changes the logo to be 72x64.

This also simplifies the CSS (removing unnecessary floats and spacing)
and fixes vertical alignment of the logo in print.